filters added to controllers

+++ app/controllers/notes_controller.rb
@@ -1,20 +1,28 @@
 class NotesController < ApplicationController
 
+    before_action :authorized
+
     def create
-        note = Note.create(note_params)
-        render json: note
+        if @user
+            note = Note.create(note_params)
+            render json: note
+        end
     end
 
     def update
-        note = Note.find(params[:id])
-        note.update(note_params)
-        render json: note
+        if @user
+            note = Note.find(params[:id])
+            note.update(note_params)
+            render json: note
+        end
     end
 
     def destroy 
-        note = Note.find(params[:id])
-        note.destroy
-        render json: note
+        if @user
+            note = Note.find(params[:id])
+            note.destroy
+            render json: note
+        end
     end
 
     private
